requirements.txt
fastapi
uvicorn
google-cloud-speech
pydub
requests
PyAudio
openai